🛢️DB : Change the type of 'answer' from Integer to Long in `Round` entity

+++ backend/src/main/java/com/toyproject/hyeonworld/api/round/infrastructure/entity/Round.java
@@ -32,7 +32,7 @@ public class Round {
   @Column(name = "game_id")
   private Long gameId;
 
-  private Integer answer;
+  private Long answer;
 
   @Column(name = "created_at")
   private LocalDateTime createdAt;
